# ouargaz-app/public/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OUARGAZ API",
    "description": "Documentation interactive des API pour la plateforme OUARGAZ (Next.js & Application Mobile).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Serveur actuel"
    }
  ],
  "paths": {
...
    "/api/notifications": {
      "get": {
        "summary": "Obtenir les notifications de l'utilisateur",
        "description": "Récupère les 25 dernières notifications destinées au rôle de l'utilisateur connecté.",
        "tags": ["Notifications"],
        "responses": {
          "200": {
            "description": "Liste de notifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notifications": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "role": { "type": "string" },
                          "title": { "type": "string" },
                          "message": { "type": "string" },
                          "link": { "type": "string" },
                          "read": { "type": "boolean" },
                          "createdAt": { "type": "string", "format": "date-time" }
                        }
                      }
                    },
                    "unread": { "type": "integer", "description": "Nombre de notifications non lues" }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Marquer une ou toutes les notifications comme lues",
        "description": "Si un `id` est fourni dans le corps, marque cette notification comme lue. Sinon, marque toutes les notifications du rôle comme lues.",
        "tags": ["Notifications"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer", "description": "ID optionnel de la notification" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Marquer toutes les notifications comme lues",
        "description": "Marque toutes les notifications du rôle connecté comme lues.",
        "tags": ["Notifications"],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/stock": {
      "get": {
        "summary": "Liste de l'état des stocks de bouteilles",
        "description": "Récupère les stocks de bouteilles enregistrés par date ou plage de dates.",
        "tags": ["Stocks"],
        "parameters": [
          { "name": "date", "in": "query", "schema": { "type": "string" }, "description": "Date YYYY-MM-DD" },
          { "name": "from", "in": "query", "schema": { "type": "string" }, "description": "Date de début YYYY-MM-DD" },
          { "name": "to", "in": "query", "schema": { "type": "string" }, "description": "Date de fin YYYY-MM-DD" }
        ],
        "responses": {
          "200": {
            "description": "Liste de stocks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/StockBouteilles" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Sauvegarder l'état des stocks pour un jour donné",
        "description": "Sauvegarde ou met à jour les stocks de bouteilles pour une date donnée.",
        "tags": ["Stocks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["date", "rows"],
                "properties": {
                  "date": { "type": "string" },
                  "rows": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["marque"],
                      "properties": {
                        "marque": { "type": "string" },
                        "pleines_12kg": { "type": "number" },
                        "pleines_6kg": { "type": "number" },
                        "pleines_3kg": { "type": "number" },
                        "defectueuses_12kg": { "type": "number" },
                        "defectueuses_6kg": { "type": "number" },
                        "defectueuses_3kg": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sauvegarde réussie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": { "description": "Date requise" },
          "403": { "description": "Journée clôturée ou accès refusé" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "StockBouteilles": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "date": { "type": "string" },
          "marque": { "type": "string" },
          "pleines_12kg": { "type": "number" },
          "pleines_6kg": { "type": "number" },
          "pleines_3kg": { "type": "number" },
          "defectueuses_12kg": { "type": "number" },
          "defectueuses_6kg": { "type": "number" },
          "defectueuses_3kg": { "type": "number" },
          "stock_pleines_kg": { "type": "number" },
          "stock_def_kg": { "type": "number" },
          "stock_cond_kg": { "type": "number" },
          "stock_cond_t": { "type": "number" }
        }
      }
    }
  }
}